Improve Synchronization between fences/Synchronsitaion?Polling/Message handling
Found that the VK_Timeout error was not in fact due to a potential bug with the Driver/Vulkan implementation itself but was in fact due to subtly mistimed fence waiting which only occurred if the framerate was above a certain value (roughly 10000 in this case) after changing the timeout period to -1 *Which is interpreted as UINT64_MAX with.IN/Via/as/Derived/Implictaed/Ascertained/Interpreted Via/ with/By Vulkan) the Framerate increasedd/improved to roughly 11200 or roughly a 8-10% improvement overall

+++ src/fakeFBO.cpp
@@ -1,6 +1,5 @@
 #include "fakeFBO.hpp"
-#include "glm/ext/vector_uint2.hpp"
-#include <ctime>
+#include "glm/ext/vector_float2.hpp"
 #include <minwindef.h>
 
 
@@ -8,7 +7,7 @@
 
 struct ITime
 {
-  glm::uvec2 xy;
+  glm::vec2 xy;
   float time;
   float stime;
 } __attribute__((aligned(16))) ;
